+++ mrv2/lib/mrvEdit/mrvEditCallbacks.cpp
@@ -32,6 +32,7 @@ namespace fs = std::filesystem;
 #include "mrvDraw/Annotation.h"
 
 #include "mrvNetwork/mrvTCP.h"
+#include "mrvNetwork/mrvInsertData.h"
 
 #include "mrvPanels/mrvPanelsCallbacks.h"
 
@@ -1731,8 +1732,8 @@ namespace mrv
         tcp->unlock();
     }
 
-    void edit_insert_clip(
-        const std::vector<tl::timeline::InsertData>& inserts, ViewerUI* ui)
+    void edit_insert_clip_annotations(
+        const std::vector<mrv::InsertData>& inserts, ViewerUI* ui)
     {
         auto player = ui->uiView->getTimelinePlayer();
         if (!player)
@@ -1752,18 +1753,18 @@ namespace mrv
         const auto& tracks = stack->children();
         for (const auto& insert : inserts)
         {
-            const int oldIndex = getIndex(insert.composable);
-            const int oldTrackIndex = getIndex(insert.composable->parent());
+            const int oldIndex = insert.oldIndex;
+            const int oldTrackIndex = insert.oldTrackIndex;
+            if (oldIndex < 0 || oldTrackIndex < 0 || insert.trackIndex < 0 ||
+                insert.trackIndex >= tracks.size())
+                continue;
+
             if (auto track = otio::dynamic_retainer_cast<otio::Track>(
                     stack->children()[oldTrackIndex]))
             {
                 if (track->kind() != otio::Track::Kind::video)
                     continue;
             }
-
-            if (oldIndex < 0 || oldTrackIndex < 0 || insert.trackIndex < 0 ||
-                insert.trackIndex >= tracks.size())
-                continue;
 
             int insertIndex = insert.insertIndex;
             if (oldTrackIndex == insert.trackIndex && oldIndex < insertIndex)
@@ -1809,6 +1810,72 @@ namespace mrv
                 }
             }
         }
+
+        ui->uiTimeline->redraw();
+    }
+
+    void
+    edit_insert_clip(const std::vector<mrv::InsertData>& inserts, ViewerUI* ui)
+    {
+        auto player = ui->uiView->getTimelinePlayer();
+        if (!player)
+            return;
+
+        std::vector<tl::timeline::InsertData> insertData;
+        const auto& timeline = player->getTimeline();
+        const auto& stack = timeline->tracks();
+        const auto& tracks = stack->children();
+        for (const auto& insert : inserts)
+        {
+            const int oldIndex = insert.oldIndex;
+            const int oldTrackIndex = insert.oldTrackIndex;
+
+            if (auto track = otio::dynamic_retainer_cast<otio::Track>(
+                    tracks[oldTrackIndex]))
+            {
+                if (auto child = track->children()[oldIndex])
+                {
+                    auto item = otio::dynamic_retainer_cast<otio::Item>(child);
+                    if (!item)
+                        continue;
+
+                    timeline::InsertData data;
+                    data.composable = child;
+                    data.trackIndex = insert.trackIndex;
+                    data.insertIndex = insert.insertIndex;
+                    insertData.push_back(data);
+                }
+            }
+        }
+
+        auto otioTimeline = tl::timeline::insert(timeline, insertData);
+        player->player()->getTimeline()->setTimeline(otioTimeline);
+
+        edit_insert_clip_annotations(inserts, ui);
+    }
+
+    void edit_insert_clip_annotations(
+        const std::vector<tl::timeline::InsertData>& inserts, ViewerUI* ui)
+    {
+        auto player = ui->uiView->getTimelinePlayer();
+        if (!player)
+            return;
+
+        // Convert tlRender data to mrv2's one.
+        std::vector<mrv::InsertData> networkInsertData;
+        for (const auto& insert : inserts)
+        {
+            const int oldIndex = getIndex(insert.composable);
+            const int oldTrackIndex = getIndex(insert.composable->parent());
+            InsertData networkInsert;
+            networkInsert.oldIndex = oldIndex;
+            networkInsert.oldTrackIndex = oldTrackIndex;
+            networkInsert.trackIndex = insert.trackIndex;
+            networkInsert.insertIndex = insert.insertIndex;
+            networkInsertData.push_back(networkInsert);
+        }
+
+        edit_insert_clip_annotations(networkInsertData, ui);
     }
 
     EditMode editMode = EditMode::kTimeline;