+++ mrv2/docs/HISTORY.md
@@ -12,6 +12,10 @@ v0.7.9
 - Fixed go to next and previous annotations when several clips of different
   rates were present in the timeline.
 - Fixed floating timeline thumbnail not updating properly when showing an EDL.
+- Added shifting of annotations when tlRender's darby routines drag an item
+  into new place.
+- Made Fit ('f' key) in timeline viewport work on network connections.
+- Made Panning (Middle mouse drag) in timeline viewport work on network connections.
 
 
 v0.7.8